Move PloneParticipationHandler to opengever.dossier.participations.
Preparation to regroup handlers for both plone and SQL
participations.
